+++ src/base/number/index.ts
@@ -2,10 +2,14 @@ export {
   NUMBER_CHI_MAP,
   addNumber,
   getPartRatio,
+
   getThousandSeparator2,
   getThousandSeparator,
   getUnitPreviousRatio,
+
   padZero,
   random,
   range,
+
+  getRandomNumber,
 } from './number';