+++ src/base/string/random.ts
@@ -1,3 +1,15 @@
+
+/**
+ * 获取随机字符串
+ * @param {number} length 字符串长度，默认 32
+ * @returns {string} 字符串
+ * @example
+ * ```ts
+ * randomString()
+ *
+ * randomString(16)
+ * ```
+ */
 export function randomString(e = 32) {
   const t = 'ABCDEFGHJKMNPQRSTWXYZabcdefhijkmnprstwxyz2345678';
   const a = t.length;
@@ -6,4 +18,15 @@ export function randomString(e = 32) {
   return n;
 }
 
+/**
+ * 获取随机字符串
+ * @param {number} length 字符串长度，默认 32
+ * @returns {string} 字符串
+ * @example
+ * ```ts
+ * randomString()
+ *
+ * randomString(16)
+ * ```
+ */
 export const getRandomString = randomString;