+++ src/rtx/index.ts
@@ -1,1 +1,2 @@
+export { getMentionRtx } from './mention';
 export { getRtxInfo } from './rtx';